+++ app/src/main/java/de/nproth/pin/receiver/AlarmReceiver.java
@@ -3,6 +3,7 @@ package de.nproth.pin.receiver;
 import android.content.BroadcastReceiver;
 import android.content.Context;
 import android.content.Intent;
+import android.preference.PreferenceManager;
 import android.util.Log;
 
 import de.nproth.pin.pinboard.Pinboard;
@@ -20,6 +21,6 @@ public class AlarmReceiver extends BroadcastReceiver {
         Intent i = new Intent(context, PinboardService.class);
         i.setAction(PinboardService.INTENT_ACTION_WAKE_UP);
         //context.startService(i);
-        Pinboard.get(context).updateVisible(false);
+        Pinboard.get(context).setSnoozeDuration(PreferenceManager.getDefaultSharedPreferences(context).getLong(PinboardService.PREFERENCE_SNOOZE_DURATION, PinboardService.DEFAULT_SNOOZE_DURATION)).updateVisible(false);
     }
 }