+++ app/src/main/java/de/nproth/pin/receiver/NotificationJobService.java
@@ -4,6 +4,7 @@ import android.annotation.TargetApi;
 import android.app.job.JobParameters;
 import android.app.job.JobService;
 import android.content.Intent;
+import android.preference.PreferenceManager;
 
 import de.nproth.pin.pinboard.Pinboard;
 import de.nproth.pin.pinboard.PinboardService;
@@ -22,7 +23,7 @@ public class NotificationJobService extends JobService {
         Intent i = new Intent(this, PinboardService.class);
         i.setAction(PinboardService.INTENT_ACTION_WAKE_UP);
         //startService(i);
-        Pinboard.get(this).updateVisible(false);
+        Pinboard.get(this).setSnoozeDuration(PreferenceManager.getDefaultSharedPreferences(this).getLong(PinboardService.PREFERENCE_SNOOZE_DURATION, PinboardService.DEFAULT_SNOOZE_DURATION)).updateVisible(false);
         return false;//work is done at this point
     }